+++ src/ehrdata/io/zarr.py
@@ -1,15 +1,20 @@
 from __future__ import annotations
 
+import warnings
+from functools import wraps
 from pathlib import Path
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any, Literal
 
 import anndata as ad
 import zarr
 
+import ehrdata as ed
 from ehrdata._logger import logger
+from ehrdata.core.constants import EHRDATA_ZARR_ENCODING_VERSION
 from ehrdata.io._array_casting import _cast_arrays_dtype_to_float_or_str_if_nonnumeric_object, _cast_variables_to_float
 
 if TYPE_CHECKING:
+    from collections.abc import Callable, Mapping
     from os import PathLike
 
     from ehrdata import EHRData
@@ -22,6 +27,8 @@ def read_zarr(
     cast_variables_to_float: bool = True,
 ) -> EHRData:
     """Read a zarr store into an :class:`~ehrdata.EHRData` object.
+
+    Can also read :class:`~anndata.AnnData` Zarr stores. In this case, a default `.tem` field is created in the `ehrdata object`.
 
     Args:
         filename: The filename, or a Zarr storage class.
@@ -36,7 +43,6 @@ def read_zarr(
         >>> ed.io.write_zarr("mimic_2.zarr", edata)
         >>> edata_from_zarr = ed.io.read_zarr("mimic_2.zarr")
     """
-    import ehrdata as ed
     from ehrdata import EHRData
 
     if isinstance(filename, Path):
@@ -44,7 +50,29 @@ def read_zarr(
 
     f = filename if isinstance(filename, zarr.Group) else zarr.open(filename, mode="r")
 
-    dictionary_for_init = {k: ad.io.read_elem(f[k]) for k, v in dict(f).items() if not k.startswith("raw.")}
+    if "encoding-type" not in f.attrs:
+        err = "The zarr store does not contain an encoding-type attribute."
+        raise ValueError(err)
+
+    if f.attrs["encoding-type"] == "ehrdata":
+        if "anndata" in f:
+            dictionary_for_init = {
+                k: ad.io.read_elem(f["anndata"][k]) for k, v in dict(f["anndata"]).items() if not k.startswith("raw.")
+            }
+        else:
+            err = "The zarr store does not contain the 'anndata' group."
+            raise ValueError(err)
+        if "tem" in f:
+            dictionary_for_init["tem"] = ad.io.read_elem(f["tem"])
+        else:
+            warnings.warn("The zarr store does not contain the 'tem' group.", stacklevel=2)
+
+    elif f.attrs["encoding-type"] == "anndata":
+        dictionary_for_init = {k: ad.io.read_elem(f[k]) for k, v in dict(f).items() if not k.startswith("raw.")}
+
+    else:
+        err = f"Unkown encoding-type '{f.attrs['encoding-type']}'."
+        raise ValueError(err)
 
     edata = EHRData(**dictionary_for_init)
 
@@ -62,23 +90,33 @@ def read_zarr(
     return edata
 
 
+def _allow_write_nullable_strings[T, **P](f: Callable[P, T]) -> Callable[P, T]:
+    @wraps(f)
+    def wrapped(*args: P.args, **kwargs: P.kwargs):
+        with ad.settings.override(allow_write_nullable_strings=True):
+            return f(*args, **kwargs)
+
+    return wrapped
+
+
+@_allow_write_nullable_strings
 def write_zarr(
     edata: EHRData,
     filename: str | Path,
     *,
-    chunks: bool | int | tuple[int, ...] | None = None,
+    chunks: Literal["auto" | "ehrdata_auto"] = "auto",
     convert_strings_to_categoricals: bool = True,
 ) -> None:
     """Write :class:`~ehrdata.EHRData` objects to disk.
 
-    To write to a `.zarr` file, `X` and `layers` cannot be written as  `object` dtype.
+    To write to a `.zarr` file, `X`, and `layers` cannot be written as `object` dtype.
     If any of these fields is of `object` dtype, it this function will attempt to cast it to a numeric dtype; if this fails, the field will be casted to a `str` dtype.
 
     Args:
         edata: Central data object.
         filename: Name of the output file, can also be prefixed with relative or absolute path to save the file to.
-        chunks: Chunk shape, passed to :meth:`zarr.Group.create_dataset` for `Zarr` version 2, or to :meth:`zarr.Group.create_array` for `Zarr` version 3.
-        convert_strings_to_categoricals: Convert columns of `str` dtype in `.obs` and `.var` to `categorical` dtype.
+        chunks: Specify strategy of how data should be chunked. For simplicity, currently only 2 options are available: `"auto"` will write the data with :func:`~anndata.io.write_elem`'s default settings. `"ehrdata_auto"` will write the data chunked (and sharded) based on a heuristic that loosely speaking writes slightly smaller chunks.
+        convert_strings_to_categoricals: Convert columns of `str` dtype in `.obs` and `.var` and `.tem` to `categorical` dtype.
 
     Examples:
         >>> import ehrdata as ed
@@ -86,13 +124,52 @@ def write_zarr(
         >>> ed.io.write_zarr("mimic_2.zarr", edata)
     """
     filename = Path(filename)
-
     edata = _cast_arrays_dtype_to_float_or_str_if_nonnumeric_object(edata)
 
-    ad.AnnData(edata).write_zarr(
-        filename,
-        chunks=chunks,
-        convert_strings_to_categoricals=convert_strings_to_categoricals,
-    )
-    f = zarr.open(filename, mode="a")
-    ad.io.write_elem(f, "tem", edata.tem)
+    store = zarr.open_group(filename, mode="a", use_consolidated=False, zarr_format=3)
+
+    adata = ad.AnnData(edata)
+
+    if convert_strings_to_categoricals:
+        adata.strings_to_categoricals(adata.obs)
+        adata.strings_to_categoricals(adata.var)
+        adata.strings_to_categoricals(edata.tem)
+
+    # write_sharded this is a slightly modified version from https://anndata.readthedocs.io/en/stable/tutorials/zarr-v3.html
+    # write_sharded is intended as a future blueprint of implementing better chunking defaults for ehrdata based based on real usecases
+    def write_sharded(group: zarr.Group, adata: ad.AnnData):
+        def callback(
+            func: ad.experimental.Write,
+            g: zarr.Group,
+            k: str,
+            elem: ad.typing.RWAble,
+            dataset_kwargs: Mapping[str, Any],
+            iospec: ad.experimental.IOSpec,
+        ):
+            if iospec.encoding_type in {"array"} and not isinstance(elem, list):
+                dataset_kwargs = {
+                    "shards": tuple(int(2 ** (16 / len(elem.shape))) for _ in elem.shape),
+                    **dataset_kwargs,
+                }
+                dataset_kwargs["chunks"] = tuple(i // 2 for i in dataset_kwargs["shards"])
+            elif iospec.encoding_type in {"csr_matrix", "csc_matrix"}:
+                dataset_kwargs = {"shards": (2**16,), "chunks": (2**8,), **dataset_kwargs}
+            func(g, k, elem, dataset_kwargs=dataset_kwargs)
+
+        return ad.experimental.write_dispatched(group, "/", adata, callback=callback)
+
+    if chunks == "auto":
+        ad.io.write_elem(store, "anndata", adata)
+    elif chunks == "ehrdata_auto":
+        anndata_group = store.create_group("anndata")
+        write_sharded(anndata_group, adata)
+    else:
+        err = (
+            f"chunks={chunks} is not implemented. Currently, only chunks='auto' and chunks='ehrdata_auto' is supported."
+        )
+        raise NotImplementedError(err)
+
+    ad.io.write_elem(store, "tem", edata.tem)
+
+    store.attrs["encoding-version"] = EHRDATA_ZARR_ENCODING_VERSION
+    store.attrs["encoding-type"] = "ehrdata"